bot/requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.3
beautifulsoup4==4.13.3
pytz==2025.1

# asyncpg==0.30.0
# aiohappyeyeballs==2.4.6
# aiohttp==3.11.12
# aiosignal==1.3.2
# annotated-types==0.7.0
# asgiref==3.8.1
# attrs==25.1.0
# certifi==2025.1.31
# charset-normalizer==3.4.1
# frozenlist==1.5.0
# idna==3.10
# magic-filter==1.0.12
# multidict==6.1.0
# propcache==0.2.1
psycopg2-binary==2.9.10
# pydantic==2.10.6
# pydantic_core==2.27.2
# python-dateutil==2.9.0.post0
# python-mimeparse==2.0.0
# six==1.17.0
# sqlparse==0.5.3
# typing_extensions==4.12.2
# tzdata==2025.1
# urllib3==2.3.0
# yarl==1.18.3